Add model module(#3)
Enable Test(#3)

Add travis.yml(#3)

Add chmod +x gradlew(#3)
