+++ .travis.yml
@@ -1,0 +1,7 @@
+language: java
+
+before_install:
+ - chmod +x gradlew  
+
+script:
+  - ./gradlew test
