+++ core/src/mmorihiro/jeweledoor/view/BasicView.kt
@@ -1,4 +1,4 @@
-package mmorihiro.jeweledoor
+package mmorihiro.jeweledoor.view
 
 import com.badlogic.gdx.graphics.Texture
 import com.badlogic.gdx.scenes.scene2d.Actor
@@ -8,8 +8,9 @@ import com.badlogic.gdx.utils.viewport.FitViewport
 import ktx.actors.centerPosition
 import ktx.actors.plus
 import ktx.assets.asset
+import mmorihiro.jeweledoor.model.Values
 
-class View : Stage(FitViewport(Values.width, Values.height)) {
+class BasicView : Stage(FitViewport(Values.width, Values.height)) {
     private var bullets: List<Image> = listOf()
 
     val backGround = object : Actor() {
@@ -25,7 +26,7 @@ class View : Stage(FitViewport(Values.width, Values.height)) {
             Image(asset<Texture>("bullet.png")).apply {
                 bullets += this
                 centerPosition(Values.width, Values.height)
-                this@View + this
+                this@BasicView + this
             }
 
     fun removeBullet(bullet: Image) {
