+++ desktop/src/mmorihiro/jeweledoor/desktop/DesktopLauncher.kt
@@ -2,14 +2,15 @@ package mmorihiro.jeweledoor.desktop
 
 import com.badlogic.gdx.backends.lwjgl.LwjglApplication
 import com.badlogic.gdx.backends.lwjgl.LwjglApplicationConfiguration
-import mmorihiro.jeweledoor.MainListener
+import mmorihiro.jeweledoor.controller.MainListener
+import mmorihiro.jeweledoor.model.Values
 
 object DesktopLauncher {
     @JvmStatic fun main(arg: Array<String>) {
         val config = LwjglApplicationConfiguration().apply {
             backgroundFPS = 60
-            width = 288
-            height = 480
+            width = Values.width.toInt()
+            height = Values.height.toInt()
             resizable = false
         }
