+++ model/src/main/mmorihiro/jeweledoor/model/Values.kt
@@ -1,0 +1,8 @@
+package mmorihiro.jeweledoor.model
+
+object Values {
+    const val width = 288f
+    const val height = 480f
+    const val centerX = width / 2
+    const val centerY = height / 2
+}
