+++ settings.gradle
@@ -1,1 +1,1 @@
-include 'desktop', 'core'
+include 'desktop', 'core', 'model'
